Take algorithms out of GCPointerMap
Also, remove the part where we adjust for offsets within boxed value
types.
